pyproject.toml
[tool.poetry]
name = "hd-lookism-bot"
version = "0.1.0"
description = "HD | Lookism - Telegram looksmax coaching bot"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.0.0"
sqlmodel = "^0.0.19"
httpx = "^0.27.0"
yookassa = "^2.1.0"
aioredis = "^2.0.1"
asyncpg = "^0.29.0"
opencv-python-headless = "^4.8.0"
pydantic = "^2.0.0"
python-multipart = "^0.0.6"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
numpy = "^1.24.0"
python-dotenv = "^1.0.0"
aiosqlite = "^0.21.0"
greenlet = "^3.2.3"
apscheduler = "^3.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
